requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Core Libraries
numpy==1.26.3
Pillow==10.2.0
opencv-python==4.9.0.80

# Document Processing
pdfplumber==0.11.0
PyMuPDF==1.23.8
pdf2image==1.16.3

# OCR Engines
easyocr==1.7.0
pytesseract==0.3.10

# AI/Vision APIs
google-genai>=0.2.0

# ML/AI Libraries
scikit-learn==1.4.0
scikit-image==0.25.2
sentence-transformers==2.2.2
